Refactor budgetswitch
