+++ client/src/components/Layout.tsx
@@ -1,18 +1,14 @@
 import { Outlet } from "react-router-dom";
 import Navbar from "./navbar/Navbar";
-import { useReducer } from "react";
-import { BudgetContext, BudgetDispatchContext, BudgetReducer } from "./navbar/BudgetSwitch/BudgetContext";
+import { BudgetStateProvider } from "./shared/BudgetContextProvider";
 
 const Layout = () => {
-  const [budget, dispatch] = useReducer(BudgetReducer, false);
   return (
     <div className="content">
-      <BudgetContext.Provider value={budget}>
-        <BudgetDispatchContext.Provider value={dispatch}>
-          <Navbar isLoggedIn={true} isAdmin={true} isLoading={false} racename={"tour"} currentStageLink={"/"} />
-          <div className="pageContainer">{<Outlet />}</div>
-        </BudgetDispatchContext.Provider>
-      </BudgetContext.Provider>
+      <BudgetStateProvider>
+        <Navbar isLoggedIn={true} isAdmin={true} isLoading={false} racename={"tour"} currentStageLink={"/"} />
+        <div className="pageContainer">{<Outlet />}</div>
+      </BudgetStateProvider>
     </div >
   );
 };
